# api/src/openapi.yml
openapi: 3.0.3
info:
  title: Offline-First Chat Application API
  description: REST API for an offline-first chat application with event-based synchronization and configurable conflict resolution
  version: 1.0.0
servers:
  - url: https://api.chatapp.com/v1
    description: Production server
paths:
...
  /groups:
    get:
      summary: Get all available chat groups
      tags: [Groups]
      responses:
        '200':
          description: List of available chat groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatGroup'
                  syncTimestamp:
                    type: string
                    format: date-time
    post:
      summary: Create a new chat group
      tags: [Groups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                description:
                  type: string
                  maxLength: 500
                localId:
                  type: string
                  description: Client-generated temporary ID for optimistic updates
              required: [name]
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatGroup'
        '409':
          description: Group name already exists
  /groups/{groupId}:
    get:
      summary: Get a specific chat group
      tags: [Groups]
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Group details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatGroup'
        '404':
          description: Group not found
    
    put:
      summary: Update a chat group
      tags: [Groups]
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                description:
                  type: string
                  maxLength: 500
              required: [name]
      responses:
        '200':
          description: Group updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatGroup'
        '403':
          description: Only group creator can update the group
        '404':
          description: Group not found
    
    delete:
      summary: Delete a chat group
      tags: [Groups]
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Group deleted successfully
        '403':
          description: Only group creator can delete the group
        '404':
          description: Group not found

  # Group Membership
  /groups/{groupId}/members:
    get:
      summary: Get current members of a group
      tags: [Membership]
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current group members
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      $ref: '#/components/schemas/GroupMember'
                  syncTimestamp:
                    type: string
                    format: date-time

    post:
      summary: Join a group
      tags: [Membership]
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                localId:
                  type: string
                  description: Client-generated temporary ID for optimistic updates
      responses:
        '201':
          description: Successfully joined group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipEvent'
        '409':
          description: Already a member of this group
  /groups/{groupId}/members/{userId}:
    delete:
      summary: Remove a user from group (leave or remove)
      tags: [Membership]
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Successfully left/removed from group
        '404':
          description: User not in group

  # Membership Event Synchronization
...
  /groups/{groupId}/membershipEvents:
    get:
      summary: Get membership events for a specific group since timestamp
      tags: [Synchronization]
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: since
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: Get events since this timestamp
      responses:
        '200':
          description: Membership events since specified time
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/MembershipEvent'
                  syncTimestamp:
                    type: string
                    format: date-time

  # Messages
  /groups/{groupId}/messages:
    get:
      summary: Get messages for a group
      tags: [Messages]
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: Get messages since this timestamp
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: Messages in the group
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  syncTimestamp:
                    type: string
                    format: date-time
                  hasMore:
                    type: boolean

    post:
      summary: Send a message to a group
      tags: [Messages]
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  maxLength: 4000
                messageType:
                  type: string
                  enum: [text, system]
                  default: text
                localId:
                  type: string
                  description: Client-generated temporary ID for optimistic updates
              required: [content]
      responses:
        '201':
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          description: Not a member of this group

  # Message Synchronization
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin-level JWT token required
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ChatGroup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entityType:
          type: string
          enum: [GROUP]
        name:
          type: string
        description:
          type: string
        createdBy:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isActive:
          type: boolean
    GroupMember:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        joinedAt:
          type: string
          format: date-time
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        username:
          type: string
        content:
          type: string
        messageType:
          type: string
          enum: [text, system]
        createdAt:
          type: string
          format: date-time
        syncStatus:
          type: string
          enum: [synced, pending, failed]
...
    MembershipEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        action:
          type: string
          enum: [JOIN, LEAVE, REMOVE]
        performedBy:
          type: string
          format: uuid
          description: User who performed the action (for REMOVE actions)
        timestamp:
          type: string
          format: date-time
...